refactor: change type from Integer to Long

+++ src/main/java/br/com/ccm/api/bugmonitor/command/notion/outputs/attribute/UniqueId.java
@@ -2,6 +2,6 @@ package br.com.ccm.api.bugmonitor.command.notion.outputs.attribute;
 
 public record UniqueId(
         String prefix,
-        Integer number
+        Long number
 ) {
 }
